refactor: 제한사항에 맞게 move 기능 수정
